+++ src/transupp.c
@@ -5,6 +5,7 @@
  * Copyright (C) 1997-2019, Thomas G. Lane, Guido Vollbeding.
  * libjpeg-turbo Modifications:
  * Copyright (C) 2010, 2017, 2021-2022, 2024, 2026, D. R. Commander.
+ * Copyright (C) 2026, Ricardo M. Ferreira.
  * For conditions of distribution and use, see the accompanying README.ijg
  * file.
  *
@@ -1184,6 +1185,67 @@ do_rot_270(j_decompress_ptr srcinfo, j_compress_ptr dstinfo,
 
 
 LOCAL(void)
+do_roll(j_decompress_ptr srcinfo, j_compress_ptr dstinfo,
+        JDIMENSION x_shift_offset, JDIMENSION y_shift_offset,
+        jvirt_barray_ptr *src_coef_arrays, jvirt_barray_ptr *dst_coef_arrays)
+/* Roll (shift with wraparound.)  x_shift_offset and y_shift_offset specify the
+ * shift amounts in iMCUs.
+ */
+{
+  JDIMENSION MCU_cols, MCU_rows, comp_width, comp_height;
+  JDIMENSION src_blk_x, src_blk_y, dst_blk_x, dst_blk_y;
+  JDIMENSION x_shift_blocks, y_shift_blocks;
+  int ci, offset_y;
+  JBLOCKARRAY src_buffer, dst_buffer;
+  jpeg_component_info *compptr;
+
+  MCU_cols = srcinfo->output_width /
+             (dstinfo->max_h_samp_factor * dstinfo_min_DCT_h_scaled_size);
+  MCU_rows = srcinfo->output_height /
+             (dstinfo->max_v_samp_factor * dstinfo_min_DCT_v_scaled_size);
+
+  for (ci = 0; ci < dstinfo->num_components; ci++) {
+    compptr = dstinfo->comp_info + ci;
+    comp_width = MCU_cols * compptr->h_samp_factor;
+    comp_height = MCU_rows * compptr->v_samp_factor;
+    x_shift_blocks = x_shift_offset * compptr->h_samp_factor;
+    y_shift_blocks = y_shift_offset * compptr->v_samp_factor;
+    for (dst_blk_y = 0; dst_blk_y < compptr->height_in_blocks;
+         dst_blk_y += compptr->v_samp_factor) {
+      dst_buffer = (*srcinfo->mem->access_virt_barray)
+        ((j_common_ptr)srcinfo, dst_coef_arrays[ci], dst_blk_y,
+         (JDIMENSION)compptr->v_samp_factor, TRUE);
+      for (offset_y = 0; offset_y < compptr->v_samp_factor; offset_y++) {
+        for (dst_blk_x = 0; dst_blk_x < compptr->width_in_blocks;
+             dst_blk_x++) {
+          /* Calculate source block position with wraparound */
+          if (dst_blk_y < y_shift_blocks) {
+            src_blk_y = comp_height - y_shift_blocks + dst_blk_y;
+          } else {
+            src_blk_y = dst_blk_y - y_shift_blocks;
+          }
+          if (dst_blk_x < x_shift_blocks) {
+            src_blk_x = comp_width - x_shift_blocks + dst_blk_x;
+          } else {
+            src_blk_x = dst_blk_x - x_shift_blocks;
+          }
+          if (dst_blk_y >= comp_height)
+            src_blk_y = dst_blk_y;
+          if (dst_blk_x >= comp_width)
+            src_blk_x = dst_blk_x;
+          src_buffer = (*srcinfo->mem->access_virt_barray)
+            ((j_common_ptr)srcinfo, src_coef_arrays[ci], src_blk_y,
+             (JDIMENSION)compptr->v_samp_factor, FALSE);
+          jcopy_block_row(src_buffer[offset_y] + src_blk_x,
+                          dst_buffer[offset_y] + dst_blk_x, (JDIMENSION)1);
+        }
+      }
+    }
+  }
+}
+
+
+LOCAL(void)
 do_rot_180(j_decompress_ptr srcinfo, j_compress_ptr dstinfo,
            JDIMENSION x_crop_offset, JDIMENSION y_crop_offset,
            jvirt_barray_ptr *src_coef_arrays,
@@ -1508,9 +1570,11 @@ LOCAL(void)
 trim_right_edge(jpeg_transform_info *info, JDIMENSION full_width)
 {
   JDIMENSION MCU_cols;
+  JDIMENSION x_crop_offset = info->transform == JXFORM_ROLL ?
+                             0 : info->x_crop_offset;
 
   MCU_cols = info->output_width / info->iMCU_sample_width;
-  if (MCU_cols > 0 && info->x_crop_offset + MCU_cols ==
+  if (MCU_cols > 0 && x_crop_offset + MCU_cols ==
       full_width / info->iMCU_sample_width)
     info->output_width = MCU_cols * info->iMCU_sample_width;
 }
@@ -1519,9 +1583,11 @@ LOCAL(void)
 trim_bottom_edge(jpeg_transform_info *info, JDIMENSION full_height)
 {
   JDIMENSION MCU_rows;
+  JDIMENSION y_crop_offset = info->transform == JXFORM_ROLL ?
+                             0 : info->y_crop_offset;
 
   MCU_rows = info->output_height / info->iMCU_sample_height;
-  if (MCU_rows > 0 && info->y_crop_offset + MCU_rows ==
+  if (MCU_rows > 0 && y_crop_offset + MCU_rows ==
       full_height / info->iMCU_sample_height)
     info->output_height = MCU_rows * info->iMCU_sample_height;
 }
@@ -1587,7 +1653,7 @@ jtransform_request_workspace(j_decompress_ptr srcinfo,
   if (info->perfect &&
       !jtransform_perfect_transform(srcinfo->output_width,
                                     srcinfo->output_height, MCU_width,
-                                    MCU_height, info->transform))
+                                    MCU_height, info))
     return FALSE;
 
   /* If there is only one output component, force the iMCU size to be 1;
@@ -1639,7 +1705,10 @@ jtransform_request_workspace(j_decompress_ptr srcinfo,
     if (info->crop_width_set == JCROP_UNSET) {
       if (info->crop_xoffset >= info->output_width)
         ERREXIT(srcinfo, JERR_BAD_CROP_SPEC);
-      info->crop_width = info->output_width - info->crop_xoffset;
+      if (info->transform == JXFORM_ROLL)
+        info->crop_width = 0;
+      else
+        info->crop_width = info->output_width - info->crop_xoffset;
     } else {
       /* Check for crop extension */
       if (info->crop_width > info->output_width) {
@@ -1660,7 +1729,10 @@ jtransform_request_workspace(j_decompress_ptr srcinfo,
     if (info->crop_height_set == JCROP_UNSET) {
       if (info->crop_yoffset >= info->output_height)
         ERREXIT(srcinfo, JERR_BAD_CROP_SPEC);
-      info->crop_height = info->output_height - info->crop_yoffset;
+      if (info->transform == JXFORM_ROLL)
+        info->crop_height = 0;
+      else
+        info->crop_height = info->output_height - info->crop_yoffset;
     } else {
       /* Check for crop extension */
       if (info->crop_height > info->output_height) {
@@ -1678,7 +1750,7 @@ jtransform_request_workspace(j_decompress_ptr srcinfo,
           ERREXIT(srcinfo, JERR_BAD_CROP_SPEC);
       }
     }
-    /* Convert negative crop offsets into regular offsets */
+    /* Convert negative crop/roll offsets into regular offsets */
     if (info->crop_xoffset_set != JCROP_NEG)
       xoffset = info->crop_xoffset;
     else if (info->crop_width > info->output_width) /* crop extension */
@@ -1753,13 +1825,13 @@ jtransform_request_workspace(j_decompress_ptr srcinfo,
       if (info->crop_width_set == JCROP_FORCE ||
           info->crop_width > info->output_width)
         info->output_width = info->crop_width;
-      else
+      else if (info->transform != JXFORM_ROLL)
         info->output_width =
           info->crop_width + (xoffset % info->iMCU_sample_width);
       if (info->crop_height_set == JCROP_FORCE ||
           info->crop_height > info->output_height)
         info->output_height = info->crop_height;
-      else
+      else if (info->transform != JXFORM_ROLL)
         info->output_height =
           info->crop_height + (yoffset % info->iMCU_sample_height);
     }
@@ -1837,6 +1909,16 @@ jtransform_request_workspace(j_decompress_ptr srcinfo,
   case JXFORM_WIPE:
     break;
   case JXFORM_DROP:
+    break;
+  case JXFORM_ROLL:
+    if (info->trim) {
+      if (info->crop_xoffset)
+        trim_right_edge(info, srcinfo->output_width);
+      if (info->crop_yoffset)
+        trim_bottom_edge(info, srcinfo->output_height);
+    }
+    /* Need workspace arrays having same dimensions as source image. */
+    need_workspace = TRUE;
     break;
   }
 
@@ -2295,6 +2377,10 @@ jtransform_execute_transform(j_decompress_ptr srcinfo, j_compress_ptr dstinfo,
               src_coef_arrays, info->drop_ptr, info->drop_coef_arrays,
               info->drop_width, info->drop_height);
     break;
+  case JXFORM_ROLL:
+    do_roll(srcinfo, dstinfo, info->x_crop_offset, info->y_crop_offset,
+            src_coef_arrays, dst_coef_arrays);
+    break;
   }
 }
 
@@ -2322,11 +2408,11 @@ jtransform_execute_transform(j_decompress_ptr srcinfo, j_compress_ptr dstinfo,
 GLOBAL(boolean)
 jtransform_perfect_transform(JDIMENSION image_width, JDIMENSION image_height,
                              int MCU_width, int MCU_height,
-                             JXFORM_CODE transform)
+                             jpeg_transform_info *info)
 {
   boolean result = TRUE; /* initialize TRUE */
 
-  switch (transform) {
+  switch (info->transform) {
   case JXFORM_FLIP_H:
   case JXFORM_ROT_270:
     if (image_width % (JDIMENSION)MCU_width)
@@ -2342,6 +2428,12 @@ jtransform_perfect_transform(JDIMENSION image_width, JDIMENSION image_height,
     if (image_width % (JDIMENSION)MCU_width)
       result = FALSE;
     if (image_height % (JDIMENSION)MCU_height)
+      result = FALSE;
+    break;
+  case JXFORM_ROLL:
+    if (image_width % (JDIMENSION)MCU_width && info->crop_xoffset)
+      result = FALSE;
+    if (image_height % (JDIMENSION)MCU_height && info->crop_yoffset)
       result = FALSE;
     break;
   default: